move auto-materialize rule impls into own module (#21386)
## Summary & Motivation

I wanted to include `AutoMaterializePolicy` as a field in a pydantic
class, but pydantic didn't like that some of its attributes had quoted
types. This allows importing the auto_materialize_rule module from the
auto_materialize_policy module, outside of a TYPE_CHECKING block.

## How I Tested These Changes
