+++ python_modules/dagster-graphql/dagster_graphql/schema/auto_materialize_policy.py
@@ -4,7 +4,7 @@ from dagster._core.definitions.auto_materialize_policy import (
     AutoMaterializePolicy,
     AutoMaterializePolicyType,
 )
-from dagster._core.definitions.auto_materialize_rule import (
+from dagster._core.definitions.auto_materialize_rule_impls import (
     AutoMaterializeDecisionType,
     DiscardOnMaxMaterializationsExceededRule,
 )
